Chore: convert to TypeScript
