+++ .npmignore
@@ -1,0 +1,3 @@
+tsconfig.json
+src
+*.tgz
