+++ package.json
@@ -2,7 +2,8 @@
   "name": "pdjr-skplugin-mqttgw",
   "version": "2.1.0",
   "description": "Exchange data between MQTT and Signal K",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "signalk-plugin-enabled-by-default": true,
   "directories": {
   },
@@ -12,7 +13,11 @@
     "signalk-liblog": "github:pdjr-signalk/signalk-liblog"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\""
+    "publish": "yq -o=json -I=2 resources/openApi.yaml > resources/openApi.json",
+    "test": "echo \"Error: no test specified\"",
+    "build": "tsc",
+    "prepare": "tsc",
+    "pack": "npm pack"
   },
   "repository": {
     "type": "git",
